refactor: move system message builder to separate module
